feat(engine): 添加音频系统重构与瓦片地图渲染优化
- 实现了音频系统的多后端支持（OpenAL、SDL3、XAudio2）
- 添加了音频设备管理与版本查询功能
- 重构了瓦片地图渲染器，优化顶点数据结构与UV坐标计算
- 更新了图层类型转换方法，提高类型安全性
- 添加了Win32输入驱动与XInput游戏手柄支持
- 集成了DirectX 12运行时库链接
- 修复了图块碰撞形状的浮点数精度问题
- 优化了瓦片地图解析与序列化功能

+++ src/engine/tilemap/core/Tile.h
@@ -36,7 +36,7 @@ enum class CollisionShapeType {
 struct CollisionShape {
     CollisionShapeType type;
     // 形状数据
-    std::vector<std::pair<int, int>> points;  // 相对坐标
+    std::vector<std::pair<float, float>> points;  // 相对坐标 (浮点数)
 };
 
 // ============================================================================
